% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{list_hai_datasets}
\alias{list_hai_datasets}
\title{List packaged HAI datasets (CSV)}
\usage{
list_hai_datasets()
}
\value{
Character vector like c("hai_table1","hai_table2")
}
\description{
List packaged HAI datasets (CSV)
}
