% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summarise_hai}
\alias{summarise_hai}
\title{Summarise HAI burden}
\usage{
summarise_hai(
  df,
  by = c("infection_type", "sample"),
  metric = c("cases", "deaths", "daly")
)
}
\arguments{
\item{df}{A data frame with columns infection_type, sample, cases, deaths, daly.}

\item{by}{One of "infection_type","sample".}

\item{metric}{One of "cases","deaths","daly".}
}
\value{
A tibble with columns `group` and `value`.
}
\description{
Summarise HAI burden
}
